Finished from regitser to homeview jump

+++ Code/Tunix/app/src/main/java/tunix/model/Admin.java
@@ -2,7 +2,7 @@ package tunix.model;
 import tunix.dto.enums.Role;
 
 public class Admin extends Account {
-    public Admin(int id, String username, String email, Role accountStatus) {
+    public Admin(Long id, String username, String email, Role accountStatus) {
         super(id, username, email, accountStatus);
     }
 }
